refatorar função LerControles para outro arquivo
