+++ controles.h
@@ -1,0 +1,8 @@
+#ifndef CONTROLES_H
+#define CONTROLES_H
+
+#include "entidade.h"
+
+void LerControles (Entidade* jogador, Entidade* monstros, int nDeMonstros);
+
+#endif
