+++ main.c
@@ -9,6 +9,7 @@
 #include "sprite.h"
 #include "mapa.h"
 #include "entidade.h"
+#include "controles.h"
 
 void DesenharStatus(int vida, int nivel, int pontuacao) {
   DrawRectangle(0, 0, LARGURA, ALTURA_STATUS, BLACK);
@@ -29,7 +30,6 @@ void IniciarMapaTeste(Mapa* m) {
     for (int j = 0; j < 24; ++j)
       m->tiles[i][j] = GRAMA;
 
-
   MapaGerarRetangulos(m);
 
   m->textura = LoadTexture("assets/tileset.png");
@@ -40,64 +40,6 @@ void IniciarMapaTeste(Mapa* m) {
   for (int i = 0; i < m->nDeMonstros; ++i) {
     m->posicaoMonstros[i].x = (rand() % 24) * 32;
     m->posicaoMonstros[i].y = ((rand() % 16) * 32) + ALTURA_STATUS;
-  }
-}
-
-void LerControles (Entidade* jogador, Entidade* monstros, int nDeMonstros) {
-  bool n,s,l,o;
-  n = IsKeyDown(KEY_W);
-  s = IsKeyDown(KEY_S);
-  l = IsKeyDown(KEY_D) && ! (n || s);
-  o = IsKeyDown(KEY_A) && ! (n || s);
-
-  if (jogador->efeitos & ATAQUE) {
-    jogador->deslocamento = (Vector2){ 0, 0 };
-    jogador->spriteAtual = ATACANDO;
-  } else {
-    jogador->deslocamento = (Vector2){
-      (n || s) ? 0 : (l - o) * VEL_JOGADOR,
-      (l || o) ? 0 : (s - n) * VEL_JOGADOR
-    };
-
-    jogador->spriteAtual = (jogador->deslocamento.x != 0 || jogador->deslocamento.y != 0) ? CAMINHANDO : PARADO;
-
-    if (n && !(l || o))
-      jogador->orientacao = NORTE;
-    else if (s && !(l || o))
-      jogador->orientacao = SUL;
-    else if (l && !(n || s))
-      jogador->orientacao = LESTE;
-    else if (o && !(n || s))
-      jogador->orientacao = OESTE;
-  }  
-
-  if (IsKeyPressed(KEY_T) && !(jogador->efeitos & ATAQUE)) {
-    jogador->efeitos = ATAQUE;
-    jogador->spriteAtual = ATACANDO;
-    Rectangle hitscan = {
-      jogador->hitbox.x, jogador->hitbox.y,
-      jogador->hitbox.height, jogador->hitbox.height
-    };
-
-    switch (jogador->orientacao) {
-    case NORTE:
-      hitscan.y -= jogador->hitbox.height;
-      break;
-    case SUL:
-      hitscan.y += jogador->hitbox.height;
-      break;
-    case LESTE: hitscan.x += jogador->hitbox.width;
-      break;
-    case OESTE: hitscan.x -= jogador->hitbox.width;
-      break;
-    }
-
-    for (int i = 0; i < nDeMonstros; ++i) {
-      if (CheckCollisionRecs(monstros[i].hitbox, hitscan)) {
-	monstros[i].efeitos |= DANO;
-	--monstros[i].vida;
-      }
-    }
   }
 }
